We deactivate  the option to delete lang  when changing  the view

+++ app/src/main/java/com/example/applanguagepgm/MainActivity.kt
@@ -116,9 +116,11 @@ class MainActivity : AppCompatActivity(),OnClickListener,MainAux {
     }
 
     override fun onDeleteLang(languageEntity: LanguageEntity) {
+        if(!isView){
         doAsync {
              LanguageApplication.database.languageDao().deleteLanguage(languageEntity)
             uiThread { mAdapter.deleteLang(languageEntity) }
+        }
         }
     }
